backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
google-generativeai
pydantic
pydantic-settings
